+++ node_interop/lib/fs.dart
@@ -76,7 +76,7 @@ abstract class FS {
   external void readlink(path,
       [optionsOrCallback, callback(error, linkString)]);
   external dynamic readlinkSync(path, [options]);
-  external void readSync(int fd, buffer, int offset, int length, int position);
+  external int readSync(int fd, buffer, int offset, int length, int position);
   external void realpath(path,
       [optionsOrCallback, void callback(error, resolvedPath)]);
   // TODO: realpath.native(path[, options], callback)
@@ -117,7 +117,7 @@ abstract class FS {
 
   /// See official documentation on all possible argument combinations:
   /// - https://nodejs.org/api/fs.html#fs_fs_writesync_fd_buffer_offset_length_position
-  external void writeSync(int fd, data, [arg1, arg2, arg3]);
+  external int writeSync(int fd, data, [arg1, arg2, arg3]);
 }
 
 @JS()