+++ node_io/lib/src/file_system_entity.dart
@@ -85,10 +85,14 @@ abstract class FileSystemEntity implements io.FileSystemEntity {
   @override
   Stream<io.FileSystemEvent> watch(
       {int events: io.FileSystemEvent.all, bool recursive: false}) {
+    // TODO: implement watch
     throw new UnimplementedError();
   }
 }
 
+/// A FileStat object represents the result of calling the POSIX stat() function
+/// on a file system object.  It is an immutable object, representing the
+/// snapshotted values returned by the stat() call.
 class FileStat implements io.FileStat {
   @override
   final DateTime changed;
@@ -138,6 +142,12 @@ class FileStat implements io.FileStat {
     );
   }
 
+  /// Asynchronously calls the operating system's stat() function on [path].
+  ///
+  /// Returns a Future which completes with a [FileStat] object containing
+  /// the data returned by stat(). If the call fails, completes the future with a
+  /// [FileStat] object with `.type` set to FileSystemEntityType.notFound and
+  /// the other fields invalid.
   static Future<FileStat> stat(String path) {
     var completer = new Completer<FileStat>();
 
@@ -151,13 +161,18 @@ class FileStat implements io.FileStat {
     }
 
     var jsCallback = js.allowInterop(callback);
-    fs.stat(path, jsCallback);
+    fs.lstat(path, jsCallback);
     return completer.future;
   }
 
+  /// Calls the operating system's stat() function on [path].
+  ///
+  /// Returns a [FileStat] object containing the data returned by stat().
+  /// If the call fails, returns a [FileStat] object with .type set to
+  /// FileSystemEntityType.notFound and the other fields invalid.
   static FileStat statSync(String path) {
     try {
-      return new FileStat._fromNodeStats(fs.statSync(path));
+      return new FileStat._fromNodeStats(fs.lstatSync(path));
     } catch (_) {
       return new FileStat._internalNotFound();
     }