+++ node_io/lib/src/link.dart
@@ -3,12 +3,16 @@
 
 import 'dart:async';
 import 'dart:io' as io;
+import 'dart:js' as js;
 import 'dart:typed_data';
 
+import 'package:node_interop/fs.dart';
 import 'package:node_interop/path.dart' as nodePath;
 
+import 'directory.dart';
 import 'file_system_entity.dart';
 
+/// Link objects are references to filesystem links.
 class Link extends FileSystemEntity implements io.Link {
   @override
   final String path;
@@ -16,6 +20,7 @@ class Link extends FileSystemEntity implements io.Link {
   Link(this.path);
 
   factory Link.fromRawPath(Uint8List rawPath) {
+    // TODO: implement fromRawPath
     throw new UnimplementedError();
   }
 
@@ -47,61 +52,111 @@ class Link extends FileSystemEntity implements io.Link {
 
   @override
   Future<Link> create(String target, {bool recursive: false}) {
-    // TODO: implement create
-    throw new UnimplementedError();
+    if (recursive) {
+      throw new UnsupportedError("Recursive flag not supported by Node.js");
+    }
+
+    final completer = new Completer<Link>();
+    void cb([err]) {
+      if (err != null) {
+        completer.completeError(err);
+      } else {
+        completer.complete(this);
+      }
+    }
+
+    final jsCallback = js.allowInterop(cb);
+    fs.symlink(target, path, jsCallback);
+    return completer.future;
   }
 
   @override
   void createSync(String target, {bool recursive: false}) {
-    // TODO: implement createSync
-    throw new UnimplementedError();
+    if (recursive) {
+      throw new UnsupportedError("Recursive flag not supported by Node.js");
+    }
+    fs.symlinkSync(target, path);
   }
 
   @override
-  Future<FileSystemEntity> delete({bool recursive: false}) {
-    // TODO: implement delete
-    throw new UnimplementedError();
+  Future<Link> delete({bool recursive: false}) {
+    if (recursive) {
+      return Future.error(
+          UnsupportedError('Recursive flag is not supported by Node.js'));
+    }
+    final Completer<Link> completer = new Completer<Link>();
+    void callback(err) {
+      if (err != null) {
+        completer.completeError(err);
+      } else {
+        completer.complete(this);
+      }
+    }
+
+    final jsCallback = js.allowInterop(callback);
+    fs.unlink(_absolutePath, jsCallback);
+    return completer.future;
   }
 
   @override
   void deleteSync({bool recursive: false}) {
-    // TODO: implement deleteSync
-    throw new UnimplementedError();
+    if (recursive) {
+      throw UnsupportedError('Recursive flag is not supported by Node.js');
+    }
+    fs.unlinkSync(_absolutePath);
   }
 
   @override
   Future<Link> rename(String newPath) {
-    // TODO: implement rename
-    throw new UnimplementedError();
+    final completer = new Completer<Link>();
+    void cb(err) {
+      if (err != null) {
+        completer.completeError(err);
+      } else {
+        completer.complete(new Link(newPath));
+      }
+    }
+
+    final jsCallback = js.allowInterop(cb);
+    fs.rename(path, newPath, jsCallback);
+    return completer.future;
   }
 
   @override
   Link renameSync(String newPath) {
-    // TODO: implement renameSync
-    throw new UnimplementedError();
+    fs.renameSync(path, newPath);
+    return new Link(newPath);
   }
 
   @override
   Future<String> target() {
-    // TODO: implement target
-    throw new UnimplementedError();
+    final completer = new Completer<String>();
+    void cb(err, String target) {
+      if (err != null) {
+        completer.completeError(err);
+      } else {
+        completer.complete(target);
+      }
+    }
+
+    final jsCallback = js.allowInterop(cb);
+    fs.readlink(path, jsCallback);
+    return completer.future;
   }
 
   @override
   String targetSync() {
-    // TODO: implement targetSync
-    throw new UnimplementedError();
+    return fs.readlinkSync(path);
   }
 
   @override
   Future<Link> update(String target) {
-    // TODO: implement update
-    throw new UnimplementedError();
+    return delete().then((link) => link.create(target));
   }
 
   @override
   void updateSync(String target) {
-    // TODO: implement updateSync
-    throw new UnimplementedError();
+    deleteSync();
+    createSync(target);
   }
 }